+++ infrastructure-detekt/src/main/kotlin/internal/String.kt
@@ -1,0 +1,3 @@
+package com.redmadrobot.build.detekt.internal
+
+internal fun String.capitalized() = this.replaceFirstChar { it.uppercaseChar() }